refactor remote sync related code
